Adding visualization for Ship class
