+++ src/battleship/__init__.py
@@ -1,0 +1,7 @@
+"""
+Import the main classes in battleship
+"""
+
+from battleship.missile import Missile
+from battleship.radar import Radar
+from battleship.ship import Ship
