+++ src/battleship/params.py
@@ -1,0 +1,7 @@
+"""
+Helpful parameters that may be used by all classes within battleship
+"""
+
+# Constants
+FULL_FIGURE_HEIGHT = 700  
+FULL_FIGURE_WIDTH = 1100
